Add an initial plugin proposed by ChatGPT.
[[[Prompt]]]

Write a babel plugin that transforms function definifions using const/let into those that uses function declarations. For example,

```javascript
const f = (x) => {
  return x * x;
};
```

Should be transformed as:

```javascript
function f(x) {
  return x * x;
}
```

[[[Answer]]]

Here is an example of a Babel plugin that would transform const/let function definitions into function declarations:

```
const f = (x) => {
  return x * x;
};
```
```
function f(x) {
  return x * x;
}
```
```
const myBabelPlugin = {
  visitor: {
    // This visitor will be applied to all `VariableDeclaration` nodes with `const` or `let` declarations
    VariableDeclaration(path) {
      if (path.node.kind === "const" || path.node.kind === "let") {
        // For each `VariableDeclarator` node in the declaration, we check if it's an arrow function
        path.node.declarations.forEach((declarator) => {
          if (declarator.init && declarator.init.type === "ArrowFunctionExpression") {
            // If it is, we create a new `FunctionDeclaration` node with the same name and function body
            const func = t.functionDeclaration(
              declarator.id,
              declarator.init.params,
              declarator.init.body,
              declarator.init.generator,
              declarator.init.async
            );

            // And replace the original `VariableDeclarator` node with the new `FunctionDeclaration` node
            path.replaceWith(func);
          }
        });
      }
    }
  }
};
```

In this example, the plugin uses Babel's `traverse` API to visit all `VariableDeclaration` nodes in the AST. For each `VariableDeclaration` node with a `const` or `let` kind, the plugin checks if it contains any `VariableDeclarator` nodes that are arrow functions. If it does, it replaces the `VariableDeclarator` node with a `FunctionDeclaration` node with the same name and function body.