Try to fix the bug
[[[Prompt]]]

So here is the new debug output:

```
$ yarn build
Converting const declaration of function with name 'anonymous' to function declaration.
Converting const declaration of function with name 'anonymous' to function declaration.
Successfully compiled 1 file with Babel (217ms).
```

Next step?

[[[Answer]]]

It looks like the if block that checks for the presence of the id property is not being executed. This is because path.node.id is not guaranteed to be truthy. In order to fix this, we need to change the condition to check for the existence of the id property, and then check if its name property equals "this". Here is how the updated code would look:

```
const t = require("@babel/types");

module.exports = function({ types: t }) {
  return {
    visitor: {
      VariableDeclaration(path) {
        // If the original arrow function uses this, we bind the correct this value to the function
        if (path.node.id && path.node.id.name === "this") {
          console.log(`Converting const declaration to bound function expression.`);
          path.node.declarations.forEach((decl) => {
            if (t.isArrowFunctionExpression(decl.init)) {
              decl.init.body = t.callExpression(
                t.bindExpression(
                  decl.init.body,
                  t.thisExpression()
                ),
                decl.init.params
              );
            }
          });
        }
        // Otherwise, we can simply convert the const declaration to a function declaration
        else {
          // Use the name of the identifier if it exists, otherwise use a default string
          const name = path.node.id ? path.node.id.name : "anonymous";
          console.log(`Converting const declaration of function with name '${name}' to function declaration.`);
          path.node.declarations.forEach((decl) => {
            if (t.isArrowFunctionExpression(decl.init)) {
              path.replaceWith(
                t.functionDeclaration(
                  decl.id,
                  decl.init.params,
                  decl.init.body
                )
              );
            }
          });
        }
      }
    }
  };
};
```

Please try this updated code and let me know if it fixes the issue.