Try to fix the bind-this bug
[[[Prompt]]]

The produced JavaScript code is still:

```javascript
function f(x) {
  return x * x;
}
function g(x) {
  return this.factor * x;
}
```

which is semantically different from the input.

Debug output is:

```
$ yarn build
Converting const declaration of function with name 'anonymous' to function declaration.
Converting const declaration of function with name 'anonymous' to function declaration.
Successfully compiled 1 file with Babel (220ms).
```

Next step?

[[[Answer]]]

It looks like the problem is that we're using the wrong test to determine whether to bind this to the function. In the current code, we're only binding this if the VariableDeclaration node has an id property with a name equal to "this". But that's not the correct test -- we should be binding this if the original arrow function uses this in its body.

To fix this, we can add a check for the presence of this in the body of the arrow function. Here's how that would look:

```
const t = require("@babel/types");

module.exports = function({ types: t }) {
  return {
    visitor: {
      VariableDeclaration(path) {
        path.node.declarations.forEach((decl) => {
          if (t.isArrowFunctionExpression(decl.init)) {
            // Check if the original arrow function uses `this` in its body
            if (t.isThisExpression(decl.init.body)) {
              console.log(`Converting const declaration of function with name '${decl.id.name}' to bound function expression.`);
              decl.init.body = t.callExpression(
                t.bindExpression(
                  decl.init.body,
                  t.thisExpression()
                ),
                decl.init.params
              );
            } else {
              console.log(`Converting const declaration of function with name '${decl.id.name}' to function declaration.`);
              path.replaceWith(
                t.functionDeclaration(
                  decl.id,
                  decl.init.params,
                  decl.init.body
                )
              );
            }
          }
        });
      }
    }
  };
};
```

With this change, the code should produce the correct output. Let me know if that works for you.